# backend/src/swagger.yaml
openapi: 3.0.0
info:
  title: Smart School LMS API
  description: |
    Comprehensive REST API for Smart School Learning Management System.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Base URL
    Development: `http://localhost:5000/api/v1`
    
  version: 1.0.0
  contact:
    name: Smart School Team
    email: support@smartschool.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.smartschool.com/api/v1
    description: Production server
...
paths:
...
  /modules:
    get:
      tags: [Modules]
      summary: Get all modules
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, PENDING_APPROVAL, APPROVED, PUBLISHED, REJECTED, ARCHIVED]
        - name: teacherId
          in: query
          schema:
            type: string
        - name: subjectId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of modules
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      modules:
                        type: array
                        items:
                          $ref: '#/components/schemas/Module'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags: [Modules]
      summary: Create new module
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModule'
      responses:
        '201':
          description: Module created successfully
        '403':
          description: Only teachers can create modules
...
  /topics:
    get:
      tags: [Topics]
      summary: Get all topics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of topics

    post:
      tags: [Topics]
      summary: Create new topic
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, moduleId]
              properties:
                title:
                  type: string
                description:
                  type: string
                moduleId:
                  type: string
                orderIndex:
                  type: integer
      responses:
        '201':
          description: Topic created successfully
...
  /lessons:
    post:
      tags: [Lessons]
      summary: Create new lesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, topicId, type]
              properties:
                title:
                  type: string
                description:
                  type: string
                topicId:
                  type: string
                type:
                  type: string
                  enum: [VIDEO, YOUTUBE_LIVE, TEXT, PDF, QUIZ, ASSIGNMENT, EXTERNAL_LINK]
                content:
                  type: string
                videoUrl:
                  type: string
                duration:
                  type: integer
                orderIndex:
                  type: integer
      responses:
        '201':
          description: Lesson created successfully
...
  /enrollments:
    get:
      tags: [Enrollments]
      summary: Get all enrollments
      security:
        - bearerAuth: []
      parameters:
        - name: moduleId
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of enrollments

    post:
      tags: [Enrollments]
      summary: Enroll student in module
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moduleId, studentId]
              properties:
                moduleId:
                  type: string
                studentId:
                  type: string
      responses:
        '201':
          description: Enrollment created successfully
        '409':
          description: Already enrolled
...
  /exams:
    get:
      tags: [Exams]
      summary: Get all exams
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [UPCOMING, ACTIVE, COMPLETED, CANCELLED]
        - name: type
          in: query
          schema:
            type: string
        - name: subjectId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of exams

    post:
      tags: [Exams]
      summary: Create new exam
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExam'
      responses:
        '201':
          description: Exam created successfully
...
  /resources:
    get:
      tags: [Resources]
      summary: Get all resources
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: moduleId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of resources

    post:
      tags: [Resources]
      summary: Upload resource
      security:
        - bearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, moduleId, type]
              properties:
                title:
                  type: string
                description:
                  type: string
                moduleId:
                  type: string
                type:
                  type: string
                category:
                  type: string
                file:
                  type: string
                  format: binary
                externalUrl:
                  type: string
      responses:
        '201':
          description: Resource uploaded successfully
...
  /notices:
    get:
      tags: [Notices]
      summary: Get all notices
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of notices

    post:
      tags: [Notices]
      summary: Create notice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content, category]
              properties:
                title:
                  type: string
                content:
                  type: string
                category:
                  type: string
                  enum: [EXAM, EVENT, HOLIDAY, GENERAL]
                moduleId:
                  type: string
      responses:
        '201':
          description: Notice created successfully
...
  /admin/users:
    get:
      tags: [Admin]
      summary: Get all users (Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of users

    post:
      tags: [Admin]
      summary: Create user (Admin)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: User created successfully
...
  /admin/analytics:
    get:
      tags: [Admin, Analytics]
      summary: Get system analytics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: System analytics data
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Module:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        thumbnailUrl:
          type: string
        level:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        duration:
          type: integer
        status:
          type: string
          enum: [DRAFT, PENDING_APPROVAL, APPROVED, PUBLISHED, REJECTED, ARCHIVED]
        isFeatured:
          type: boolean
        teacherId:
          type: string
        subjectId:
          type: string
        classId:
          type: string
        createdAt:
          type: string
          format: date-time
    CreateModule:
      type: object
      required: [title, subjectId, classId]
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: string
        classId:
          type: string
        level:
          type: string
        duration:
          type: integer
        thumbnailUrl:
          type: string
...
    CreateExam:
      type: object
      required: [title, subjectId, type, startTime, endTime, duration, totalMarks]
      properties:
        title:
          type: string
        description:
          type: string
        subjectId:
          type: string
        classId:
          type: string
        type:
          type: string
          enum: [MIDTERM, FINAL, QUIZ, ASSIGNMENT, PROJECT]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        duration:
          type: integer
        totalMarks:
          type: number
        passingMarks:
          type: number
        questions:
          type: array
          items:
            type: object
    Pagination:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
...